requirements.txt
altair==6.0.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==3.2.2
blinker==1.9.0
cachetools==6.2.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0.post1
exceptiongroup==1.3.0
Faker==37.12.0
fastapi==0.120.4
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
limits==5.6.0
MarkupSafe==3.0.3
narwhals==2.14.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
plotly==6.5.0
protobuf==6.33.2
psutil==7.1.3
pyarrow==22.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
streamlit==1.52.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.1
